requirements.txt
pandas               # Data manipulation and analysis; used for cleaning and transforming CSV data
sqlalchemy           # SQL toolkit and ORM for connecting Python to databases
psycopg2-binary      # PostgreSQL adapter for Python; enables connection to Postgres DB
jupyter              # Interactive coding notebooks for analysis and documentation
matplotlib           # Core plotting library; used for creating visual charts
seaborn              # Statistical data visualization built on matplotlib for trend analysis
python-dotenv        # Loads environment variables from a .env file